Extract Expr related functions from MongoDbPlanner into separate file
